add another page in user interface
