+++ WebApp/userinterface/src/App.css
@@ -36,17 +36,3 @@
     transform: rotate(360deg);
   }
 }
-
-/*some css I added*/
-input[type="file"] {
-  display: inline-block;
- }
- .table {
-  max-width: 90%;
-  margin: 10px;
- }
- .table th {
-  text-align: center;
- }
- /*end of my css*/
-
