+++ WebApp/userinterface/src/First.js
@@ -1,10 +1,29 @@
-import React from "react";
+import React,{useState} from "react";
 import history from "./utils/history";
 
 function First() {
+
+  const [name,setName] = useState("ramachandran");
+
+
+  const callBtn = () =>{
+
+    setName("Pragadesh");
+
+    alert(name);
+
+    
+    }
+
+
   return (
     <>
+
       <div class="display-4 mb-1">This is the First Route</div>
+<h1>
+  { name}
+</h1>
+
       <button
         class="btn btn-info btn-block"
         type="button"
@@ -23,7 +42,18 @@ function First() {
       >
         Second
       </button>
+
+
+
+      <button
+        class="btn btn-info btn-block"
+        type="button"
+        onClick= {callBtn}>
+        Third
+      </button>
+
     </>
+
   );
 }
