+++ WebApp/userinterface/src/Nft.js
@@ -1,0 +1,1330 @@
+import './App.css';
+import {useState} from 'react';
+import web3 from './web3';
+import lottery from './nftcontract';//this line import lottery folder
+import ipfs from './ipfs';
+import storehash from './storehash';
+import { Router, Route, Switch } from "react-router-dom";
+import history from "./utils/history";
+import First from "./First";
+import Second from "./Second";
+import NewPage from './Newpage';
+
+
+function App() {
+
+  const [toaddress,setToaddress] = useState("");
+  const [tid,setId] = useState("");
+  const [turi,setUri] = useState("");
+  const [tname,setName] = useState("");
+  const [tsymbol,setSymbol] = useState("");
+  const [tokenuri,setTokenUri] = useState("");
+  var [getimageurl,setgetImage] = useState("");
+  const [tfile,setTfile] = useState("");
+  const [ipfsHash,setIpfsHash] = useState(null);
+  var [buffer,setBuffer] = useState("");
+  const [ethAddress,setEthAddress] = useState("");
+  const [blockNumber,setBlocknumber] = useState("");
+  const [transactionHash,setTransaction] = useState("");
+  const [gasUsed,setGasUsed] = useState("");
+  const [txReceipt,se] = useState("");
+
+
+  //start
+
+
+  const captureFile =(event) => {
+    event.stopPropagation()
+    event.preventDefault()
+    const file = event.target.files[0]
+    let reader = new window.FileReader()
+    reader.readAsArrayBuffer(file)
+    reader.onloadend = () => convertToBuffer(reader);    
+  };
+const convertToBuffer = async(reader) => {
+  //file is converted to a buffer for upload to IPFS
+    const buffer = await Buffer.from(reader.result);
+  //set this buffer -using es6 syntax
+    setBuffer(buffer);
+};
+const onSubmitImage = async (event) => {
+  event.preventDefault();
+ //bring in user's metamask account address
+  const accounts = await web3.eth.getAccounts();
+ 
+  console.log('Sending from Metamask account: ' + accounts[0]);
+
+  await ipfs.add(buffer, (err, ipfsHash) => {
+    console.log(err,ipfsHash);
+    setIpfsHash(ipfsHash[0].hash);
+  }) 
+}; 
+//end
+
+  
+
+   const onSubmitNFT = async (event) => {
+
+
+    //alert(localStorage.getItem('myimageuri'));
+      
+
+    event.preventDefault();
+    
+      //new write below
+
+      const accounts = await web3.eth.getAccounts();
+    
+    
+      alert("waiting for pic url");
+    
+      var ta=tname;
+      var tb=tsymbol;
+      var tc='https://ipfs.io/ipfs/'+ipfsHash;
+      var td=toaddress;
+      var te=tid;
+      var tf='https://ipfs.io/ipfs/'+ipfsHash;
+      
+      
+    
+      alert("im work ta tb tc  td te tf  "+ta+" "+tb+" "+tc+" "+td+" "+te+" "+tf);
+    
+      
+    
+      //const accounts = await  web3.eth.getAccounts();
+    
+    
+      await lottery.deploy({
+    
+      data:'60806040523480156200001157600080fd5b506040516200424538038062004245833981018060405260a08110156200003757600080fd5b8101908080516401000000008111156200005057600080fd5b820160208101848111156200006457600080fd5b81516401000000008111828201871017156200007f57600080fd5b505092919060200180516401000000008111156200009c57600080fd5b82016020810184811115620000b057600080fd5b8151640100000000811182820187101715620000cb57600080fd5b50509291906020018051640100000000811115620000e857600080fd5b82016020810184811115620000fc57600080fd5b81516401000000008111828201871017156200011757600080fd5b505060208083015160409093015191945091925090859085908290829062000165907f01ffc9a70000000000000000000000000000000000000000000000000000000090620002c1811b901c565b6200017d6380ac58cd60e01b620002c160201b60201c565b6200019563780e9d6360e01b620002c160201b60201c565b8151620001aa90600990602085019062000679565b508051620001c090600a90602084019062000679565b50620001d9635b5e139f60e01b620002c160201b60201c565b50505050620001fd620001f16200039060201b60201c565b6200039560201b60201c565b6200020d6200039060201b60201c565b600d80546001600160a01b0319166001600160a01b03928316179081905560405191169060009060008051602062004203833981519152908290a3601180546001600160a01b038084166001600160a01b031992831617909255601280549285169290911691909117905582516200028d90601090602086019062000679565b506000600f55620002a581620003e7602090811b901c565b620002b6816200039560201b60201c565b50505050506200071b565b7fffffffff0000000000000000000000000000000000000000000000000000000080821614156200035357604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601c60248201527f4552433136353a20696e76616c696420696e7465726661636520696400000000604482015290519081900360640190fd5b7fffffffff00000000000000000000000000000000000000000000000000000000166000908152602081905260409020805460ff19166001179055565b335b90565b620003b081600c6200047760201b62002f4c1790919060201c565b6040516001600160a01b038216907f6ae172837ea30b801fbfcdd4108aa1d5bf8ff775444fd70256b44e6bf3dfc3f690600090a250565b620003f76200051b60201b60201c565b6200046357604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820181905260248201527f4f776e61626c653a2063616c6c6572206973206e6f7420746865206f776e6572604482015290519081900360640190fd5b62000474816200054a60201b60201c565b50565b620004898282620005f660201b60201c565b15620004f657604080517f08c379a000000000000000000000000000000000000000000000000000000000815260206004820152601f60248201527f526f6c65733a206163636f756e7420616c72656164792068617320726f6c6500604482015290519081900360640190fd5b6001600160a01b0316600090815260209190915260409020805460ff19166001179055565b600d546000906001600160a01b03166200053b62000390602090811b901c565b6001600160a01b031614905090565b6001600160a01b038116620005ab576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401808060200182810382526026815260200180620041dd6026913960400191505060405180910390fd5b600d546040516001600160a01b038084169216906000805160206200420383398151915290600090a3600d80546001600160a01b0319166001600160a01b0392909216919091179055565b60006001600160a01b03821662000659576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401808060200182810382526022815260200180620042236022913960400191505060405180910390fd5b506001600160a01b03166000908152602091909152604090205460ff1690565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f10620006bc57805160ff1916838001178555620006ec565b82800160010185558215620006ec579182015b82811115620006ec578251825591602001919060010190620006cf565b50620006fa929150620006fe565b5090565b6200039291905b80821115620006fa576000815560010162000705565b613ab2806200072b6000396000f3fe6080604052600436106102255760003560e01c80638da5cb5b11610123578063af640d0f116100ab578063d4477b3c1161006f578063d4477b3c14610bba578063dc8e92ea14610c6a578063e564da6914610d18578063e985e9c514610dc8578063f2fde38b14610e0357610225565b8063af640d0f146109d0578063b4a99a4e146109e5578063b88d4fde146109fa578063bfb231d214610acb578063c87b56dd14610b9057610225565b806398588a2b116100f257806398588a2b1461086b57806398650275146109385780639abc83201461094d578063a22cb46514610962578063aa271e1a1461099d57610225565b80638da5cb5b146107f95780638f32d59b1461080e57806395d89b4114610823578063983b2d561461083857610225565b806342966c68116101b15780636352211e116101755780636352211e146106e757806370a0823114610711578063715018a61461074457806380482491146107595780638462151c1461077657610225565b806342966c68146105a357806345cefcec146105cd5780634f6ccce7146105e257806350655d8c1461060c57806350bb4e7f1461062157610225565b80631239ec8c116101f85780631239ec8c1461037d57806318160ddd146104bd57806323b872dd146104e45780632f745c591461052757806342842e0e1461056057610225565b806301ffc9a71461022a57806306fdde0314610272578063081812fc146102fc578063095ea7b314610342575b600080fd5b34801561023657600080fd5b5061025e6004803603602081101561024d57600080fd5b50356001600160e01b031916610e36565b604080519115158252519081900360200190f35b34801561027e57600080fd5b50610287610e55565b6040805160208082528351818301528351919283929083019185019080838360005b838110156102c15781810151838201526020016102a9565b50505050905090810190601f1680156102ee5780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b34801561030857600080fd5b506103266004803603602081101561031f57600080fd5b5035610eec565b604080516001600160a01b039092168252519081900360200190f35b34801561034e57600080fd5b5061037b6004803603604081101561036557600080fd5b506001600160a01b038135169060200135610f51565b005b34801561038957600080fd5b5061037b600480360360608110156103a057600080fd5b6001600160a01b038235169190810190604081016020820135600160201b8111156103ca57600080fd5b8201836020820111156103dc57600080fd5b803590602001918460208302840111600160201b831117156103fd57600080fd5b9190808060200260200160405190810160405280939291908181526020018383602002808284376000920191909152509295949360208101935035915050600160201b81111561044c57600080fd5b82018360208201111561045e57600080fd5b803590602001918460208302840111600160201b8311171561047f57600080fd5b91908080602002602001604051908101604052809392919081815260200183836020028082843760009201919091525092955061107f945050505050565b3480156104c957600080fd5b506104d261127f565b60408051918252519081900360200190f35b3480156104f057600080fd5b5061037b6004803603606081101561050757600080fd5b506001600160a01b03813581169160208101359091169060400135611285565b34801561053357600080fd5b506104d26004803603604081101561054a57600080fd5b506001600160a01b0381351690602001356112e4565b34801561056c57600080fd5b5061037b6004803603606081101561058357600080fd5b506001600160a01b03813581169160208101359091169060400135611367565b3480156105af57600080fd5b5061037b600480360360208110156105c657600080fd5b5035611382565b3480156105d957600080fd5b5061037b6113d7565b3480156105ee57600080fd5b506104d26004803603602081101561060557600080fd5b503561146a565b34801561061857600080fd5b506103266114d3565b34801561062d57600080fd5b5061025e6004803603606081101561064457600080fd5b6001600160a01b0382351691602081013591810190606081016040820135600160201b81111561067357600080fd5b82018360208201111561068557600080fd5b803590602001918460018302840111600160201b831117156106a657600080fd5b91908080601f0160208091040260200160405190810160405280939291908181526020018383808284376000920191909152509295506114e2945050505050565b3480156106f357600080fd5b506103266004803603602081101561070a57600080fd5b5035611550565b34801561071d57600080fd5b506104d26004803603602081101561073457600080fd5b50356001600160a01b03166115a7565b34801561075057600080fd5b5061037b611612565b61025e6004803603602081101561076f57600080fd5b50356116b8565b34801561078257600080fd5b506107a96004803603602081101561079957600080fd5b50356001600160a01b031661190e565b60408051602080825283518183015283519192839290830191858101910280838360005b838110156107e55781810151838201526020016107cd565b505050509050019250505060405180910390f35b34801561080557600080fd5b5061032661196f565b34801561081a57600080fd5b5061025e61197e565b34801561082f57600080fd5b506102876119a4565b34801561084457600080fd5b5061037b6004803603602081101561085b57600080fd5b50356001600160a01b0316611a05565b34801561087757600080fd5b5061037b600480360360a081101561088e57600080fd5b6001600160a01b0382351691602081013591810190606081016040820135600160201b8111156108bd57600080fd5b8201836020820111156108cf57600080fd5b803590602001918460018302840111600160201b831117156108f057600080fd5b91908080601f01602080910402602001604051908101604052809392919081815260200183838082843760009201919091525092955050823593505050602001351515611a57565b34801561094457600080fd5b5061037b611c47565b34801561095957600080fd5b50610287611c59565b34801561096e57600080fd5b5061037b6004803603604081101561098557600080fd5b506001600160a01b0381351690602001351515611ce7565b3480156109a957600080fd5b5061025e600480360360208110156109c057600080fd5b50356001600160a01b0316611def565b3480156109dc57600080fd5b506104d2611e02565b3480156109f157600080fd5b50610326611e08565b348015610a0657600080fd5b5061037b60048036036080811015610a1d57600080fd5b6001600160a01b03823581169260208101359091169160408201359190810190608081016060820135600160201b811115610a5757600080fd5b820183602082011115610a6957600080fd5b803590602001918460018302840111600160201b83111715610a8a57600080fd5b91908080601f016020809104026020016040519081016040528093929190818152602001838380828437600092019190915250929550611e17945050505050565b348015610ad757600080fd5b50610af560048036036020811015610aee57600080fd5b5035611e78565b6040518085815260200184815260200180602001836003811115610b1557fe5b60ff168152602001828103825284818151815260200191508051906020019080838360005b83811015610b52578181015183820152602001610b3a565b50505050905090810190601f168015610b7f5780820380516001836020036101000a031916815260200191505b509550505050505060405180910390f35b348015610b9c57600080fd5b5061028760048036036020811015610bb357600080fd5b5035611f2a565b348015610bc657600080fd5b5061037b60048036036040811015610bdd57600080fd5b810190602081018135600160201b811115610bf757600080fd5b820183602082011115610c0957600080fd5b803590602001918460208302840111600160201b83111715610c2a57600080fd5b919080806020026020016040519081016040528093929190818152602001838360200280828437600092019190915250929550505050351515905061205e565b348015610c7657600080fd5b5061037b60048036036020811015610c8d57600080fd5b810190602081018135600160201b811115610ca757600080fd5b820183602082011115610cb957600080fd5b803590602001918460208302840111600160201b83111715610cda57600080fd5b919080806020026020016040519081016040528093929190818152602001838360200280828437600092019190915250929550612282945050505050565b348015610d2457600080fd5b5061037b60048036036040811015610d3b57600080fd5b810190602081018135600160201b811115610d5557600080fd5b820183602082011115610d6757600080fd5b803590602001918460208302840111600160201b83111715610d8857600080fd5b9190808060200260200160405190810160405280939291908181526020018383602002808284376000920191909152509295505091359250612423915050565b348015610dd457600080fd5b5061025e60048036036040811015610deb57600080fd5b506001600160a01b03813581169160200135166124b3565b348015610e0f57600080fd5b5061037b60048036036020811015610e2657600080fd5b50356001600160a01b03166124e1565b6001600160e01b03191660009081526020819052604090205460ff1690565b60098054604080516020601f6002600019610100600188161502019095169490940493840181900481028201810190925282815260609390929091830182828015610ee15780601f10610eb657610100808354040283529160200191610ee1565b820191906000526020600020905b815481529060010190602001808311610ec457829003601f168201915b505050505090505b90565b6000610ef782612546565b610f3557604051600160e51b62461bcd02815260040180806020018281038252602c815260200180613911602c913960400191505060405180910390fd5b506000908152600260205260409020546001600160a01b031690565b6000610f5c82611550565b9050806001600160a01b0316836001600160a01b03161415610fb257604051600160e51b62461bcd0281526004018080602001828103825260218152602001806139b46021913960400191505060405180910390fd5b806001600160a01b0316610fc4612563565b6001600160a01b03161480610fe55750610fe581610fe0612563565b6124b3565b61102357604051600160e51b62461bcd0281526004018080602001828103825260388152602001806138146038913960400191505060405180910390fd5b60008281526002602052604080822080546001600160a01b0319166001600160a01b0387811691821790925591518593918516917f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b92591a4505050565b60005b815181101561110f576110bc8484838151811061109b57fe5b60200260200101518484815181106110af57fe5b6020026020010151611285565b6003600e60008484815181106110ce57fe5b6020026020010151815260200190815260200160002060030160006101000a81548160ff0219169083600381111561110257fe5b0217905550600101611082565b507fe0bdff057bc74b2545c13ac89393038d336f254b2245c876efa9669ddcc09b97600e60008360008151811061114257fe5b602002602001015181526020019081526020016000206002018383604051808060200180602001806020018481038452878181546001816001161561010002031660029004815260200191508054600181600116156101000203166002900480156111ee5780601f106111c3576101008083540402835291602001916111ee565b820191906000526020600020905b8154815290600101906020018083116111d157829003601f168201915b50508481038352865181528651602091820191808901910280838360005b8381101561122457818101518382015260200161120c565b50505050905001848103825285818151815260200191508051906020019060200280838360005b8381101561126357818101518382015260200161124b565b50505050905001965050505050505060405180910390a1505050565b60075490565b611296611290612563565b82612567565b6112d457604051600160e51b62461bcd0281526004018080602001828103825260318152602001806139d56031913960400191505060405180910390fd5b6112df83838361260e565b505050565b60006112ef836115a7565b821061132f57604051600160e51b62461bcd02815260040180806020018281038252602b815260200180613741602b913960400191505060405180910390fd5b6001600160a01b038316600090815260056020526040902080548390811061135357fe5b906000526020600020015490505b92915050565b6112df83838360405180602001604052806000815250611e17565b61138d611290612563565b6113cb57604051600160e51b62461bcd028152600401808060200182810382526030815260200180613a576030913960400191505060405180910390fd5b6113d48161262d565b50565b6113df61197e565b6114335760408051600160e51b62461bcd02815260206004820181905260248201527f4f776e61626c653a2063616c6c6572206973206e6f7420746865206f776e6572604482015290519081900360640190fd5b6040517ff58fef8e187ef8dfd7bef096c1ef3e4f3c54f98d95b8ad5659349b07e61204df90600090a16011546001600160a01b0316ff5b600061147461127f565b82106114b457604051600160e51b62461bcd02815260040180806020018281038252602c815260200180613a06602c913960400191505060405180910390fd5b600782815481106114c157fe5b90600052602060002001549050919050565b6011546001600160a01b031681565b60006114f46114ef612563565b611def565b61153257604051600160e51b62461bcd02815260040180806020018281038252603081526020018061389f6030913960400191505060405180910390fd5b61153c848461263f565b6115468383612660565b5060019392505050565b6000818152600160205260408120546001600160a01b03168061136157604051600160e51b62461bcd0281526004018080602001828103825260298152602001806138766029913960400191505060405180910390fd5b60006001600160a01b0382166115f157604051600160e51b62461bcd02815260040180806020018281038252602a81526020018061384c602a913960400191505060405180910390fd5b6001600160a01b0382166000908152600360205260409020611361906126c6565b61161a61197e565b61166e5760408051600160e51b62461bcd02815260206004820181905260248201527f4f776e61626c653a2063616c6c6572206973206e6f7420746865206f776e6572604482015290519081900360640190fd5b600d546040516000916001600160a01b0316907f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0908390a3600d80546001600160a01b0319169055565b6000818152600e60205260408120600101543410156117125760408051600160e51b62461bcd02815260206004820152600b6024820152600160a81b6a507269636520697373756502604482015290519081900360640190fd5b6000828152600e6020526040902060039081015460ff169081111561173357fe5b6001146117775760408051600160e51b62461bcd0281526020600482015260076024820152600160c81b664e6f2053616c6502604482015290519081900360640190fd5b6000828152600e60205261178a346126ca565b9050600061179834836126ec565b6011546040519192506001600160a01b03169082156108fc029083906000818181858888f193505050501580156117d3573d6000803e3d6000fd5b506012546040516001600160a01b039091169083156108fc029084906000818181858888f1935050505015801561180e573d6000803e3d6000fd5b505050601154611828906001600160a01b0316338461260e565b6000828152600e602090815260409182902060038101805460ff19166002908117909155835186815234948101859052606093810184815292820180546001811615610100026000190116929092049381018490527f46e70e87b058d8f3cde334fc179a61ac8acadcc4fdc26735e4195106b3c88eb3948794929390926080830190859080156118f95780601f106118ce576101008083540402835291602001916118f9565b820191906000526020600020905b8154815290600101906020018083116118dc57829003601f168201915b505094505050505060405180910390a1919050565b60606119198261272e565b80548060200260200160405190810160405280929190818152602001828054801561196357602002820191906000526020600020905b81548152602001906001019080831161194f575b50505050509050919050565b600d546001600160a01b031690565b600d546000906001600160a01b0316611995612563565b6001600160a01b031614905090565b600a8054604080516020601f6002600019610100600188161502019095169490940493840181900481028201810190925282815260609390929091830182828015610ee15780601f10610eb657610100808354040283529160200191610ee1565b611a106114ef612563565b611a4e57604051600160e51b62461bcd02815260040180806020018281038252603081526020018061389f6030913960400191505060405180910390fd5b6113d481612748565b611a626114ef612563565b611aa057604051600160e51b62461bcd02815260040180806020018281038252603081526020018061389f6030913960400191505060405180910390fd5b6028841115611ae65760408051600160e51b62461bcd0281526020600482015260076024820152600160cc1b6604f7665722034302604482015290519081900360640190fd5b60005b84811015611c3f57600f54611b0590600163ffffffff61279016565b600f8181556000918252600e602090815260408084206001018790559154835291208551611b3b92600290920191870190613644565b5060018215151415611b6b57600f546000908152600e60205260409020600301805460ff19166001179055611b88565b600f546000908152600e60205260409020600301805460ff191690555b611b9486600f5461263f565b7fadef11a3979b8ceb0573eb6ef0678134a09c23a0d94e5ea47cd18ac3a9fc0194600f54856040518083815260200180602001828103825283818151815260200191508051906020019080838360005b83811015611bfc578181015183820152602001611be4565b50505050905090810190601f168015611c295780820380516001836020036101000a031916815260200191505b50935050505060405180910390a1600101611ae9565b505050505050565b611c57611c52612563565b6127ed565b565b6010805460408051602060026001851615610100026000190190941693909304601f81018490048402820184019092528181529291830182828015611cdf5780601f10611cb457610100808354040283529160200191611cdf565b820191906000526020600020905b815481529060010190602001808311611cc257829003601f168201915b505050505081565b611cef612563565b6001600160a01b0316826001600160a01b03161415611d585760408051600160e51b62461bcd02815260206004820152601960248201527f4552433732313a20617070726f766520746f2063616c6c657200000000000000604482015290519081900360640190fd5b8060046000611d65612563565b6001600160a01b03908116825260208083019390935260409182016000908120918716808252919093529120805460ff191692151592909217909155611da9612563565b60408051841515815290516001600160a01b0392909216917f17307eab39ab6107e8899845ad3d59bd9653f200f220920489ca2b5937696c319181900360200190a35050565b6000611361600c8363ffffffff61283516565b600f5481565b6013546001600160a01b031681565b611e28611e22612563565b83612567565b611e6657604051600160e51b62461bcd0281526004018080602001828103825260318152602001806139d56031913960400191505060405180910390fd5b611e728484848461289f565b50505050565b600e6020908152600091825260409182902080546001808301546002808501805488519481161561010002600019011691909104601f810187900487028401870190975286835292959094919291830182828015611f175780601f10611eec57610100808354040283529160200191611f17565b820191906000526020600020905b815481529060010190602001808311611efa57829003601f168201915b5050506003909301549192505060ff1684565b6010805460408051602060026001851615610100026000190190941693909304601f81018490048402820184019092528181526060936113619391929091830182828015611fb95780601f10611f8e57610100808354040283529160200191611fb9565b820191906000526020600020905b815481529060010190602001808311611f9c57829003601f168201915b5050506000868152600e60209081526040918290206002908101805484516001821615610100026000190190911692909204601f810184900484028301840190945283825290945092508301828280156120545780601f1061202957610100808354040283529160200191612054565b820191906000526020600020905b81548152906001019060200180831161203757829003601f168201915b50505050506128f4565b6120696114ef612563565b6120a757604051600160e51b62461bcd02815260040180806020018281038252603081526020018061389f6030913960400191505060405180910390fd5b60005b8251811015612162576001821515141561210e576001600e60008584815181106120d057fe5b6020026020010151815260200190815260200160002060030160006101000a81548160ff0219169083600381111561210457fe5b021790555061215a565b6000600e600085848151811061212057fe5b6020026020010151815260200190815260200160002060030160006101000a81548160ff0219169083600381111561215457fe5b02179055505b6001016120aa565b507fefd56044628f374cec68c952a0580a76d23b4bd06d17bb887193f2176ba6c48b82600e60008560008151811061219657fe5b60200260200101518152602001908152602001600020600201604051808060200180602001838103835285818151815260200191508051906020019060200280838360005b838110156121f35781810151838201526020016121db565b50505050919091018481038352855460026000196101006001841615020190911604808252602090910191508590801561226e5780601f106122435761010080835404028352916020019161226e565b820191906000526020600020905b81548152906001019060200180831161225157829003601f168201915b505094505050505060405180910390a15050565b61228d6114ef612563565b6122cb57604051600160e51b62461bcd02815260040180806020018281038252603081526020018061389f6030913960400191505060405180910390fd5b60005b81518110156122fb576122f38282815181106122e657fe5b602002602001015161262d565b6001016122ce565b507f68f2536b9ff968bd2acc006b7cf7be7ca83c2f4a462c355e925bb354e5ce43d5600e60008360008151811061232e57fe5b60200260200101518152602001908152602001600020600201826040518080602001806020018381038352858181546001816001161561010002031660029004815260200191508054600181600116156101000203166002900480156123d55780601f106123aa576101008083540402835291602001916123d5565b820191906000526020600020905b8154815290600101906020018083116123b857829003601f168201915b50508381038252845181528451602091820191808701910280838360005b8381101561240b5781810151838201526020016123f3565b5050505090500194505050505060405180910390a150565b61242e6114ef612563565b61246c57604051600160e51b62461bcd02815260040180806020018281038252603081526020018061389f6030913960400191505060405180910390fd5b60005b82518110156112df5781600e600085848151811061248957fe5b6020026020010151815260200190815260200160002060010181905550808060010191505061246f565b6001600160a01b03918216600090815260046020908152604080832093909416825291909152205460ff1690565b6124e961197e565b61253d5760408051600160e51b62461bcd02815260206004820181905260248201527f4f776e61626c653a2063616c6c6572206973206e6f7420746865206f776e6572604482015290519081900360640190fd5b6113d481612930565b6000908152600160205260409020546001600160a01b0316151590565b3390565b600061257282612546565b6125b057604051600160e51b62461bcd02815260040180806020018281038252602c8152602001806137e8602c913960400191505060405180910390fd5b60006125bb83611550565b9050806001600160a01b0316846001600160a01b031614806125f65750836001600160a01b03166125eb84610eec565b6001600160a01b0316145b80612606575061260681856124b3565b949350505050565b6126198383836129d4565b6126238382612b1e565b6112df8282612c13565b6113d461263982611550565b82612c51565b6126498282612c99565b6126538282612c13565b61265c81612dd0565b5050565b61266982612546565b6126a757604051600160e51b62461bcd02815260040180806020018281038252602c81526020018061393d602c913960400191505060405180910390fd5b6000828152600b6020908152604090912082516112df92840190613644565b5490565b6000806126d8836002612e14565b90506126e5816064612e70565b9392505050565b60006126e583836040518060400160405280601e81526020017f536166654d6174683a207375627472616374696f6e206f766572666c6f770000815250612eb2565b6001600160a01b0316600090815260056020526040902090565b612759600c8263ffffffff612f4c16565b6040516001600160a01b038216907f6ae172837ea30b801fbfcdd4108aa1d5bf8ff775444fd70256b44e6bf3dfc3f690600090a250565b6000828201838110156126e55760408051600160e51b62461bcd02815260206004820152601b60248201527f536166654d6174683a206164646974696f6e206f766572666c6f770000000000604482015290519081900360640190fd5b6127fe600c8263ffffffff612fd016565b6040516001600160a01b038216907fe94479a9f7e1952cc78f2d6baab678adc1b772d936c6583def489e524cb6669290600090a250565b60006001600160a01b03821661287f57604051600160e51b62461bcd0281526004018080602001828103825260228152602001806139696022913960400191505060405180910390fd5b506001600160a01b03166000908152602091909152604090205460ff1690565b6128aa84848461260e565b6128b68484848461303a565b611e7257604051600160e51b62461bcd02815260040180806020018281038252603281526020018061376c6032913960400191505060405180910390fd5b60606126e58383604051806020016040528060008152506040518060200160405280600081525060405180602001604052806000815250613194565b6001600160a01b03811661297857604051600160e51b62461bcd02815260040180806020018281038252602681526020018061379e6026913960400191505060405180910390fd5b600d546040516001600160a01b038084169216907f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e090600090a3600d80546001600160a01b0319166001600160a01b0392909216919091179055565b826001600160a01b03166129e782611550565b6001600160a01b031614612a2f57604051600160e51b62461bcd02815260040180806020018281038252602981526020018061398b6029913960400191505060405180910390fd5b6001600160a01b038216612a7757604051600160e51b62461bcd0281526004018080602001828103825260248152602001806137c46024913960400191505060405180910390fd5b612a80816133a8565b6001600160a01b0383166000908152600360205260409020612aa1906133e3565b6001600160a01b0382166000908152600360205260409020612ac2906133fa565b60008181526001602052604080822080546001600160a01b0319166001600160a01b0386811691821790925591518493918716917fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef91a4505050565b6001600160a01b038216600090815260056020526040812054612b4890600163ffffffff6126ec16565b600083815260066020526040902054909150808214612be3576001600160a01b0384166000908152600560205260408120805484908110612b8557fe5b906000526020600020015490508060056000876001600160a01b03166001600160a01b031681526020019081526020016000208381548110612bc357fe5b600091825260208083209091019290925591825260069052604090208190555b6001600160a01b0384166000908152600560205260409020805490612c0c9060001983016136c2565b5050505050565b6001600160a01b0390911660009081526005602081815260408084208054868652600684529185208290559282526001810183559183529091200155565b612c5b8282613403565b6000818152600b6020526040902054600260001961010060018416150201909116041561265c576000818152600b6020526040812061265c916136e6565b6001600160a01b038216612cf75760408051600160e51b62461bcd02815260206004820181905260248201527f4552433732313a206d696e7420746f20746865207a65726f2061646472657373604482015290519081900360640190fd5b612d0081612546565b15612d555760408051600160e51b62461bcd02815260206004820152601c60248201527f4552433732313a20746f6b656e20616c7265616479206d696e74656400000000604482015290519081900360640190fd5b600081815260016020908152604080832080546001600160a01b0319166001600160a01b038716908117909155835260039091529020612d94906133fa565b60405181906001600160a01b038416906000907fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef908290a45050565b600780546000838152600860205260408120829055600182018355919091527fa66cc928b5edb82af9bd49922954155ab7b0942694bea4ce44661d9a8736c6880155565b600082612e2357506000611361565b82820282848281612e3057fe5b04146126e557604051600160e51b62461bcd0281526004018080602001828103825260218152602001806138f06021913960400191505060405180910390fd5b60006126e583836040518060400160405280601a81526020017f536166654d6174683a206469766973696f6e206279207a65726f00000000000081525061342f565b60008184841115612f4457604051600160e51b62461bcd0281526004018080602001828103825283818151815260200191508051906020019080838360005b83811015612f09578181015183820152602001612ef1565b50505050905090810190601f168015612f365780820380516001836020036101000a031916815260200191505b509250505060405180910390fd5b505050900390565b612f568282612835565b15612fab5760408051600160e51b62461bcd02815260206004820152601f60248201527f526f6c65733a206163636f756e7420616c72656164792068617320726f6c6500604482015290519081900360640190fd5b6001600160a01b0316600090815260209190915260409020805460ff19166001179055565b612fda8282612835565b61301857604051600160e51b62461bcd0281526004018080602001828103825260218152602001806138cf6021913960400191505060405180910390fd5b6001600160a01b0316600090815260209190915260409020805460ff19169055565b600061304e846001600160a01b0316613497565b61305a57506001612606565b6000846001600160a01b031663150b7a02613073612563565b8887876040518563ffffffff1660e01b815260040180856001600160a01b03166001600160a01b03168152602001846001600160a01b03166001600160a01b0316815260200183815260200180602001828103825283818151815260200191508051906020019080838360005b838110156130f85781810151838201526020016130e0565b50505050905090810190601f1680156131255780820380516001836020036101000a031916815260200191505b5095505050505050602060405180830381600087803b15801561314757600080fd5b505af115801561315b573d6000803e3d6000fd5b505050506040513d602081101561317157600080fd5b50516001600160e01b031916600160e11b630a85bd010214915050949350505050565b6060808690506060869050606086905060608690506060869050606081518351855187518951010101016040519080825280601f01601f1916602001820160405280156131e8576020820181803883390190505b509050806000805b88518110156132415788818151811061320557fe5b602001015160f81c60f81b83838060010194508151811061322257fe5b60200101906001600160f81b031916908160001a9053506001016131f0565b5060005b87518110156132965787818151811061325a57fe5b602001015160f81c60f81b83838060010194508151811061327757fe5b60200101906001600160f81b031916908160001a905350600101613245565b5060005b86518110156132eb578681815181106132af57fe5b602001015160f81c60f81b8383806001019450815181106132cc57fe5b60200101906001600160f81b031916908160001a90535060010161329a565b5060005b85518110156133405785818151811061330457fe5b602001015160f81c60f81b83838060010194508151811061332157fe5b60200101906001600160f81b031916908160001a9053506001016132ef565b5060005b84518110156133955784818151811061335957fe5b602001015160f81c60f81b83838060010194508151811061337657fe5b60200101906001600160f81b031916908160001a905350600101613344565b50909d9c50505050505050505050505050565b6000818152600260205260409020546001600160a01b0316156113d457600090815260026020526040902080546001600160a01b0319169055565b80546133f690600163ffffffff6126ec16565b9055565b80546001019055565b61340d82826134ce565b6134178282612b1e565b60008181526006602052604081205561265c816135a8565b6000818361348157604051600160e51b62461bcd028152602060048201818152835160248401528351909283926044909101919085019080838360008315612f09578181015183820152602001612ef1565b50600083858161348d57fe5b0495945050505050565b6000813f7fc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a47081158015906126065750141592915050565b816001600160a01b03166134e182611550565b6001600160a01b03161461352957604051600160e51b62461bcd028152600401808060200182810382526025815260200180613a326025913960400191505060405180910390fd5b613532816133a8565b6001600160a01b0382166000908152600360205260409020613553906133e3565b60008181526001602052604080822080546001600160a01b0319169055518291906001600160a01b038516907fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef908390a45050565b6007546000906135bf90600163ffffffff6126ec16565b600083815260086020526040812054600780549394509092849081106135e157fe5b9060005260206000200154905080600783815481106135fc57fe5b6000918252602080832090910192909255828152600890915260409020829055600780549061362f9060001983016136c2565b50505060009182525060086020526040812055565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f1061368557805160ff19168380011785556136b2565b828001600101855582156136b2579182015b828111156136b2578251825591602001919060010190613697565b506136be929150613726565b5090565b8154818355818111156112df576000838152602090206112df918101908301613726565b50805460018160011615610100020316600290046000825580601f1061370c57506113d4565b601f0160209004906000526020600020908101906113d491905b610ee991905b808211156136be576000815560010161372c56fe455243373231456e756d657261626c653a206f776e657220696e646578206f7574206f6620626f756e64734552433732313a207472616e7366657220746f206e6f6e20455243373231526563656976657220696d706c656d656e7465724f776e61626c653a206e6577206f776e657220697320746865207a65726f20616464726573734552433732313a207472616e7366657220746f20746865207a65726f20616464726573734552433732313a206f70657261746f7220717565727920666f72206e6f6e6578697374656e7420746f6b656e4552433732313a20617070726f76652063616c6c6572206973206e6f74206f776e6572206e6f7220617070726f76656420666f7220616c6c4552433732313a2062616c616e636520717565727920666f7220746865207a65726f20616464726573734552433732313a206f776e657220717565727920666f72206e6f6e6578697374656e7420746f6b656e4d696e746572526f6c653a2063616c6c657220646f6573206e6f74206861766520746865204d696e74657220726f6c65526f6c65733a206163636f756e7420646f6573206e6f74206861766520726f6c65536166654d6174683a206d756c7469706c69636174696f6e206f766572666c6f774552433732313a20617070726f76656420717565727920666f72206e6f6e6578697374656e7420746f6b656e4552433732314d657461646174613a2055524920736574206f66206e6f6e6578697374656e7420746f6b656e526f6c65733a206163636f756e7420697320746865207a65726f20616464726573734552433732313a207472616e73666572206f6620746f6b656e2074686174206973206e6f74206f776e4552433732313a20617070726f76616c20746f2063757272656e74206f776e65724552433732313a207472616e736665722063616c6c6572206973206e6f74206f776e6572206e6f7220617070726f766564455243373231456e756d657261626c653a20676c6f62616c20696e646578206f7574206f6620626f756e64734552433732313a206275726e206f6620746f6b656e2074686174206973206e6f74206f776e4552433732314275726e61626c653a2063616c6c6572206973206e6f74206f776e6572206e6f7220617070726f766564a165627a7a72305820835045d6d3213980a8a6c2665b91fd20f825a1096b66dca78817dfa50f020e2a00294f776e61626c653a206e6577206f776e657220697320746865207a65726f20616464726573738be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0526f6c65733a206163636f756e7420697320746865207a65726f2061646472657373',
+      arguments: [ta,tb,tc,'0xEBB8509A162bAf75A48A3e4d33e8dda28D148284','0xB324cB722bCed8A8481B0cf38d8C4a75cF0d2c44']
+      
+      })
+      .send({
+      from: accounts[0],
+      gas: 4796559,
+      gasPrice: '20000000000'
+    })
+    .then(function(newContractInstance){
+      console.log(newContractInstance.options.address) // instance with the new contract address
+       //acc=newContractInstance.options.address
+      alert(newContractInstance.options.address);
+    
+      localStorage.setItem('myData', newContractInstance.options.address);
+    
+    
+      //{()=>calling(newContractInstance.options.address)};
+    
+      //{this.calling(newContractInstance.options.address)};
+    
+      //this.setState({getting:"WORKING OR NOT"});
+    
+      //alert(this.state.getting);
+    
+    
+        
+      alert("completed");
+    
+      //alert(this.state.tname+" and "+this.state.tsymbol +" and "+this.state.tokenuri);
+    })
+    
+    
+    
+    alert(localStorage.getItem('myData'));
+    
+    var address =  localStorage.getItem('myData');
+    
+    const abi = [
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "account",
+            "type": "address"
+          }
+        ],
+        "name": "addMinter",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "to",
+            "type": "address"
+          },
+          {
+            "name": "tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "approve",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "tokenIds",
+            "type": "uint256[]"
+          }
+        ],
+        "name": "batchBurn",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "to",
+            "type": "address"
+          },
+          {
+            "name": "amountToMint",
+            "type": "uint256"
+          },
+          {
+            "name": "metaId",
+            "type": "string"
+          },
+          {
+            "name": "setPrice",
+            "type": "uint256"
+          },
+          {
+            "name": "isForSale",
+            "type": "bool"
+          }
+        ],
+        "name": "batchMint",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "giver",
+            "type": "address"
+          },
+          {
+            "name": "recipients",
+            "type": "address[]"
+          },
+          {
+            "name": "values",
+            "type": "uint256[]"
+          }
+        ],
+        "name": "batchTransfer",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "burn",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "_tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "buyThing",
+        "outputs": [
+          {
+            "name": "",
+            "type": "bool"
+          }
+        ],
+        "payable": true,
+        "stateMutability": "payable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [],
+        "name": "destroyAndSend",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "to",
+            "type": "address"
+          },
+          {
+            "name": "tokenId",
+            "type": "uint256"
+          },
+          {
+            "name": "tokenURI",
+            "type": "string"
+          }
+        ],
+        "name": "mintWithTokenURI",
+        "outputs": [
+          {
+            "name": "",
+            "type": "bool"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [],
+        "name": "renounceMinter",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [],
+        "name": "renounceOwnership",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "from",
+            "type": "address"
+          },
+          {
+            "name": "to",
+            "type": "address"
+          },
+          {
+            "name": "tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "safeTransferFrom",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "from",
+            "type": "address"
+          },
+          {
+            "name": "to",
+            "type": "address"
+          },
+          {
+            "name": "tokenId",
+            "type": "uint256"
+          },
+          {
+            "name": "_data",
+            "type": "bytes"
+          }
+        ],
+        "name": "safeTransferFrom",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "to",
+            "type": "address"
+          },
+          {
+            "name": "approved",
+            "type": "bool"
+          }
+        ],
+        "name": "setApprovalForAll",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "ids",
+            "type": "uint256[]"
+          },
+          {
+            "name": "setPrice",
+            "type": "uint256"
+          }
+        ],
+        "name": "setTokenPrice",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "ids",
+            "type": "uint256[]"
+          },
+          {
+            "name": "isEnabled",
+            "type": "bool"
+          }
+        ],
+        "name": "setTokenState",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "from",
+            "type": "address"
+          },
+          {
+            "name": "to",
+            "type": "address"
+          },
+          {
+            "name": "tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "transferFrom",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "constant": false,
+        "inputs": [
+          {
+            "name": "newOwner",
+            "type": "address"
+          }
+        ],
+        "name": "transferOwnership",
+        "outputs": [],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "function"
+      },
+      {
+        "inputs": [
+          {
+            "name": "name",
+            "type": "string"
+          },
+          {
+            "name": "symbol",
+            "type": "string"
+          },
+          {
+            "name": "uri",
+            "type": "string"
+          },
+          {
+            "name": "fee",
+            "type": "address"
+          },
+          {
+            "name": "creator",
+            "type": "address"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "nonpayable",
+        "type": "constructor"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": false,
+            "name": "error",
+            "type": "string"
+          },
+          {
+            "indexed": false,
+            "name": "tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "ErrorOut",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": false,
+            "name": "metaId",
+            "type": "string"
+          },
+          {
+            "indexed": false,
+            "name": "recipients",
+            "type": "address[]"
+          },
+          {
+            "indexed": false,
+            "name": "ids",
+            "type": "uint256[]"
+          }
+        ],
+        "name": "BatchTransfered",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": false,
+            "name": "id",
+            "type": "uint256"
+          },
+          {
+            "indexed": false,
+            "name": "metaId",
+            "type": "string"
+          }
+        ],
+        "name": "Minted",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": false,
+            "name": "metaId",
+            "type": "string"
+          },
+          {
+            "indexed": false,
+            "name": "ids",
+            "type": "uint256[]"
+          }
+        ],
+        "name": "BatchBurned",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": false,
+            "name": "ids",
+            "type": "uint256[]"
+          },
+          {
+            "indexed": false,
+            "name": "metaId",
+            "type": "string"
+          }
+        ],
+        "name": "BatchForSale",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": false,
+            "name": "tokenId",
+            "type": "uint256"
+          },
+          {
+            "indexed": false,
+            "name": "metaId",
+            "type": "string"
+          },
+          {
+            "indexed": false,
+            "name": "value",
+            "type": "uint256"
+          }
+        ],
+        "name": "Bought",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [],
+        "name": "Destroy",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": true,
+            "name": "previousOwner",
+            "type": "address"
+          },
+          {
+            "indexed": true,
+            "name": "newOwner",
+            "type": "address"
+          }
+        ],
+        "name": "OwnershipTransferred",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": true,
+            "name": "account",
+            "type": "address"
+          }
+        ],
+        "name": "MinterAdded",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": true,
+            "name": "account",
+            "type": "address"
+          }
+        ],
+        "name": "MinterRemoved",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": true,
+            "name": "from",
+            "type": "address"
+          },
+          {
+            "indexed": true,
+            "name": "to",
+            "type": "address"
+          },
+          {
+            "indexed": true,
+            "name": "tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "Transfer",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": true,
+            "name": "owner",
+            "type": "address"
+          },
+          {
+            "indexed": true,
+            "name": "approved",
+            "type": "address"
+          },
+          {
+            "indexed": true,
+            "name": "tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "Approval",
+        "type": "event"
+      },
+      {
+        "anonymous": false,
+        "inputs": [
+          {
+            "indexed": true,
+            "name": "owner",
+            "type": "address"
+          },
+          {
+            "indexed": true,
+            "name": "operator",
+            "type": "address"
+          },
+          {
+            "indexed": false,
+            "name": "approved",
+            "type": "bool"
+          }
+        ],
+        "name": "ApprovalForAll",
+        "type": "event"
+      },
+      {
+        "constant": true,
+        "inputs": [
+          {
+            "name": "owner",
+            "type": "address"
+          }
+        ],
+        "name": "balanceOf",
+        "outputs": [
+          {
+            "name": "",
+            "type": "uint256"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [],
+        "name": "baseUri",
+        "outputs": [
+          {
+            "name": "",
+            "type": "string"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [
+          {
+            "name": "tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "getApproved",
+        "outputs": [
+          {
+            "name": "",
+            "type": "address"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [],
+        "name": "id",
+        "outputs": [
+          {
+            "name": "",
+            "type": "uint256"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [
+          {
+            "name": "owner",
+            "type": "address"
+          },
+          {
+            "name": "operator",
+            "type": "address"
+          }
+        ],
+        "name": "isApprovedForAll",
+        "outputs": [
+          {
+            "name": "",
+            "type": "bool"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [
+          {
+            "name": "account",
+            "type": "address"
+          }
+        ],
+        "name": "isMinter",
+        "outputs": [
+          {
+            "name": "",
+            "type": "bool"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [],
+        "name": "isOwner",
+        "outputs": [
+          {
+            "name": "",
+            "type": "bool"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [
+          {
+            "name": "",
+            "type": "uint256"
+          }
+        ],
+        "name": "items",
+        "outputs": [
+          {
+            "name": "tokenId",
+            "type": "uint256"
+          },
+          {
+            "name": "price",
+            "type": "uint256"
+          },
+          {
+            "name": "metaId",
+            "type": "string"
+          },
+          {
+            "name": "state",
+            "type": "uint8"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [],
+        "name": "maker",
+        "outputs": [
+          {
+            "name": "",
+            "type": "address"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [],
+        "name": "name",
+        "outputs": [
+          {
+            "name": "",
+            "type": "string"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [],
+        "name": "owner",
+        "outputs": [
+          {
+            "name": "",
+            "type": "address"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [],
+        "name": "Owner",
+        "outputs": [
+          {
+            "name": "",
+            "type": "address"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [
+          {
+            "name": "tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "ownerOf",
+        "outputs": [
+          {
+            "name": "",
+            "type": "address"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [
+          {
+            "name": "interfaceId",
+            "type": "bytes4"
+          }
+        ],
+        "name": "supportsInterface",
+        "outputs": [
+          {
+            "name": "",
+            "type": "bool"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [],
+        "name": "symbol",
+        "outputs": [
+          {
+            "name": "",
+            "type": "string"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [
+          {
+            "name": "index",
+            "type": "uint256"
+          }
+        ],
+        "name": "tokenByIndex",
+        "outputs": [
+          {
+            "name": "",
+            "type": "uint256"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [
+          {
+            "name": "owner",
+            "type": "address"
+          },
+          {
+            "name": "index",
+            "type": "uint256"
+          }
+        ],
+        "name": "tokenOfOwnerByIndex",
+        "outputs": [
+          {
+            "name": "",
+            "type": "uint256"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [
+          {
+            "name": "owner",
+            "type": "address"
+          }
+        ],
+        "name": "tokensOfOwner",
+        "outputs": [
+          {
+            "name": "",
+            "type": "uint256[]"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [
+          {
+            "name": "_tokenId",
+            "type": "uint256"
+          }
+        ],
+        "name": "tokenURI",
+        "outputs": [
+          {
+            "name": "",
+            "type": "string"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      },
+      {
+        "constant": true,
+        "inputs": [],
+        "name": "totalSupply",
+        "outputs": [
+          {
+            "name": "",
+            "type": "uint256"
+          }
+        ],
+        "payable": false,
+        "stateMutability": "view",
+        "type": "function"
+      }
+    ];
+    var geta=new web3.eth.Contract(abi,address);
+    
+    
+    //let a=this.state.toaddress;
+    //let b=this.state.tokenid;
+    //let c=this.state.tokenuri;
+    
+    
+    
+    alert("im work a +b"+td+" "+te+" "+tf);
+    
+    event.preventDefault();
+    
+    //const accounts = await web3.eth.getAccounts();
+    
+    await geta.methods.mintWithTokenURI(td,te,tf).send({
+      from: accounts[0]
+      
+    });
+    
+    
+    await geta.methods.tokenURI(te).send({
+    from: accounts[0]
+    //value: this.setState({c:accounts[0]})
+    
+    });
+    
+    //this.setState({ message: 'You have been entered!' });
+    
+    //var data =await lottery.methods.tokenURI(te).call();
+    
+    //this.setState({da: String(te)})
+    
+    //this.setState({ message: 'You have been entered!' });
+    
+    alert(" and "+tid +" and "+te+"and   data   ");
+    
+    
+    alert(localStorage.getItem('myData'));
+    
+    
+    //transfer
+    
+    //await geta.methods.tokenURI(te).send({
+      //from: accounts[0]
+      //value: this.setState({c:accounts[0]})
+      
+    //});
+    
+    
+    alert("NFT NAME    "+ta+"   NFT SYMBOL    "+tb+"    TOKEN URI    "+tc+"   TOKEN ADDRESS    "+td+"    TOKEN ID   "+te+"  TOKEN URI  "+tf);
+    
+    
+    
+    
+    
+      //end above
+    
+    
+      alert("waiting for data clear");
+    
+      //document.getElementById("nameid").innerHTML="";
+      //document.getElementById("symbolid").innerHTML="";
+      //document.getElementById("addressid").innerHTML="";
+      //document.getElementById("idid").innerHTML="";
+      //document.getElementById("fileid").innerHTML="";
+    
+    
+      //this.tname.value = "";
+      //this.tsymbol.value = "";
+      //this.toaddress.value = "";
+      //this.tid.value = "";
+      //this.tfile.value = "";
+    
+    
+  };
+    
+
+  
+
+
+  return (
+    <div className="App">
+      
+<center>
+<br></br>
+<h1>DEPLOY NFT Token</h1>
+
+<form onSubmit={onSubmitImage}>
+
+
+		  <label for="images">Upload Your Image     </label>
+            <input 
+			name="tfile"
+			id="fileid"
+              type = "file"
+              onChange = {captureFile}
+			  required
+            />
+			
+             <button 
+             type="submit"> 
+             Upload Image NFT
+             </button>
+			 <br></br>
+			 <br></br>
+			 <br></br>
+	</form>
+
+
+<form onSubmit={onSubmitNFT} id="create-course-form" >
+
+<label for="name">Create Your NFT  Name    </label>
+	  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+<input
+id="nameid"
+  type='text'
+  name="tname"
+  required
+  onChange={event => setName(event.target.value)}
+  
+/>
+<br></br>
+      <br></br>
+
+	  <label for="symbol">Create Your NFT  Symbol    </label>
+	  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+
+<input
+id="symbolid"
+  type='text'
+  name="tsymbol"
+  required
+  onChange={event => setId( event.target.value)}
+  
+/>
+
+<br></br>
+<br></br>
+
+<label for="address">Enter Your Mint Address  </label>
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+
+
+<input
+id="addressid"
+  type='text'
+  name="toaddress"
+  required
+  onChange={event => setToaddress( event.target.value)}
+  
+/>
+<br></br>
+      <br></br>
+
+	  <label for="id">Create Your Unique Token-Id {' '}   </label>
+
+<input
+id="idid"
+  type='text'
+  name="tid"
+  required
+  onChange={event => setId( event.target.value)}
+  
+/>
+
+<br></br>
+<br></br>
+
+
+<button 
+             type="submit"> 
+             Upload and Create NFT
+             </button>
+
+</form>
+
+
+
+      
+<br></br>
+<br></br>
+
+
+<button
+                class="btn btn-info btn-block"
+                type="button"
+                onClick={() => {
+                  history.push("/newpage");
+                }}>
+                Get Image Pagess
+              </button>
+              <button
+                class="btn btn-info btn-block"
+                type="button"
+                onClick={() => {
+                  history.push("/second");
+                }}
+              >
+                Transfer page
+              </button>
+
+
+</center>
+
+<br></br>
+<br></br>
+  <table bordered responsive>
+                
+                <tbody>
+                  <tr>
+                    <td>Name</td>
+                    <td>{tname}</td>
+                  </tr>
+				  <tr>
+                    <td>Symbol</td>
+                    <td>{tsymbol}</td>
+                  </tr>
+				  
+				  <tr>
+                    <td>Token-Id</td>
+                    <td>{tid}</td>
+                  </tr>
+				  <tr>
+                    <td>Image-URL</td>
+                    <td>https://ipfs.io/ipfs/{ipfsHash}</td>
+                  </tr>
+				  <tr>
+                    <td>Image-View</td>
+
+                    <td><img src={localStorage.getItem('myimageuri')}  alt={'C - language'} /> 			
+					</td>
+                  </tr>         
+                </tbody>
+
+            </table>
+
+			
+
+
+
+            <Router history={history}>
+          <Switch>
+            <Route path="/" exact>
+              <div class="display-4 mb-1">Choose a route to go to</div>
+              
+            </Route>
+            <Route path="/newpage">
+              <NewPage />
+            </Route>
+            <Route path="/second">
+              <Second />
+            </Route>
+          </Switch>
+        </Router>
+
+	  
+      </div>      
+  );
+}
+
+export default App;
