--- WebApp/userinterface/src/demoview.js
@@ -1,19 +1,0 @@
-import React from 'react';
-//import ReactDOM from 'react-dom';
-import { BrowserRouter, Route,Switch,Link} from 'react-router-dom';
-
-//import Example from './example';
-class Demo extends React.Component{
-
-    state={}
-  render() {
-    return (
-      <div>
-        <h1>Landing page</h1>
-        
-      </div>
-    );
-  }
-}
-
-export default Demo;
