--- WebApp/userinterface/src/example.js
@@ -1,19 +1,0 @@
-import React from 'react';
-//import ReactDOM from 'react-dom';
-import { BrowserRouter, Route,Switch,Link} from 'react-router-dom';
-
-class Example extends React.Component{
-
-    render(){
-        return(
-
-            <div>
-                <h1>Welcome to Redirect Page Pragadesh</h1>
-                
-            </div>
-        )
-
-    }
-}
-
-export default Example;
