+++ WebApp/userinterface/src/index.js
@@ -1,31 +1,5 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-//import { BrowserRouter } from 'react-router-dom';
-//import Home from './Home';
-
-import { BrowserRouter, Route,Switch} from 'react-router-dom';
-import landingPage from './demoview';
-
-
-
+import React from "react";
+import ReactDOM from "react-dom";
+import App from "./App";
 
 ReactDOM.render(<App />, document.getElementById("root"));
-
-//ReactDOM.render(
-  //<BrowserRouter>
-
-                
-    //<App />
-  //</BrowserRouter>,
-  //document.getElementById('root')
-//);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-//reportWebVitals();
-
-//ReactDOM.render(routing,document.getElementById('root'));
